add relations between User and record models

+++ routes/modules/records.js
@@ -9,15 +9,17 @@ router.get('/new', (req, res) => {
 })
 
 router.post('/', (req, res) => {
+  const userId = req.user._id
   const { name, date, amount, category } = req.body
-  return Record.create({ name, date, amount })
+  return Record.create({ name, date, amount, userId })
     .then(() => res.redirect('/'))
     .catch(err => console.log(err))
 })
 
 router.get('/:id/edit', (req, res) => {
-  const id = req.params.id
-  return Record.findById(id)
+  const userId = req.user._id
+  const _id = req.params.id
+  return Record.findById({ _id, userId })
     .lean()
     .then((record) => {
       record.date = moment(record.date).format('YYYY-MM-DD')
@@ -27,15 +29,17 @@ router.get('/:id/edit', (req, res) => {
 })
 
 router.put('/:id', (req, res) => {
+  const userId = req.user._id
   const _id = req.params.id
-  return Record.findByIdAndUpdate({ _id }, { ...req.body })
+  return Record.findByIdAndUpdate({ _id, userId }, { ...req.body })
     .then(() => res.redirect('/'))
     .catch(err => console.log(err))
 })
 
 router.delete('/:id', (req, res) => {
+  const userId = req.user._id
   const _id = req.params.id
-  return Record.findByIdAndDelete({ _id })
+  return Record.findByIdAndDelete({ _id, userId })
     .then(() => res.redirect('/'))
     .catch(err => console.log(err))
 })
